[docs] aws 파일 분리 [docs] app.py에서 aws 파일로 분리
